moving code to more appropriate places
